Moved configuration into authenticator package from cmd
We want to dry out the cmd module as much as we can so that adding of
tests becomes easier.
